refactor: context for tests, context for actual app
