+++ backend/src/context.ts
@@ -1,22 +1,11 @@
 import { PrismaClient } from "@prisma/client";
-import { mockDeep, DeepMockProxy } from "jest-mock-extended";
 
 export type Context = {
     prisma: PrismaClient;
 };
 
-export type MockContext = {
-    prisma: DeepMockProxy<PrismaClient>;
-};
-
 export const createContext = (): Context => {
     return {
         prisma: new PrismaClient(),
-    };
-};
-
-export const createMockContext = (): MockContext => {
-    return {
-        prisma: mockDeep<PrismaClient>(),
     };
 };
